# openapi.yaml
openapi: 3.1.0
info:
  description: "This is the REST API documentation of the Spring Petclinic backend.\
    \ If authentication is enabled, use admin/admin when calling the APIs"
  termsOfService: https://github.com/spring-petclinic/petclinic-rest/blob/master/terms.txt
  title: REST Petclinic Backend API
  version: "1.0"
servers:
- description: Generated server url
  url: http://localhost
paths:
...
  /api/specialties:
    get:
      operationId: listSpecialties
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpecialtyDto"
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - specialty-rest-controller
    post:
      operationId: addSpecialty
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialtyDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - specialty-rest-controller
  /api/specialties/{specialtyId}:
    delete:
      operationId: deleteSpecialty
      parameters:
      - in: path
        name: specialtyId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - specialty-rest-controller
    get:
      operationId: getSpecialty
      parameters:
      - in: path
        name: specialtyId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SpecialtyDto"
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - specialty-rest-controller
    put:
      operationId: updateSpecialty
      parameters:
      - in: path
        name: specialtyId
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialtyDto"
        required: true
      responses:
        "200":
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                type: string
          description: Not Found
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Internal Server Error
      tags:
      - specialty-rest-controller
...
components:
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
...
    SpecialtyDto:
      type: object
      properties:
        description:
          type: string
          description: Free-text knowledge used by the chatbot to decide if this specialty
            fits a symptom.
          example: "Symptoms: limping... Guidance: keep the pet calm..."
          maxLength: 4000
          minLength: 0
        id:
          type: integer
          format: int32
          description: The ID of the specialty.
          example: 1
          minimum: 0
          readOnly: true
        name:
          type: string
          description: The name of the specialty.
          example: radiology
          maxLength: 80
          minLength: 1
      required:
      - id
      - name
...